requirements.txt
# === Core Machine Learning ===
torch==2.0.1
torchvision==0.15.2
torchdata==0.6.1
timm>=0.9.0
scikit-learn>=1.2.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3

# === Vision / Image Processing ===
Pillow==11.0.0
matplotlib>=3.7.0

# === LLM & RAG Tools ===
openai>=1.0.0
llama-index>=0.9.0
langchain>=0.1.0
tiktoken>=0.4.0

# === Graph Neural Networks (GNN) ===
dgl==1.1.3
networkx==3.4.2

# === Reinforcement Learning (RL) ===
gym>=0.26.0
stable-baselines3>=2.0.0

# === UI / Backend ===
streamlit>=1.22.0
fastapi>=0.95.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
flask>=2.2.5
flask-cors>=3.0.10
requests==2.32.3

# === Development / Utilities ===
python-dotenv>=1.0.0
watchdog>=3.0.0

# === Testing ===
pytest>=7.3.0

# === Dependency fixes ===
numpy==1.26.4
pandas==2.2.3
pyyaml==6.0.2
pydantic==2.11.4
packaging==25.0
